+++ app/routes/components/Select/index.tsx
@@ -1,12 +1,14 @@
-import { ComponentTemplate } from "src/components/ComponentTemplate";
-import { Section } from "src/components/Section";
+import { ComponentTemplate } from "@/components/ComponentTemplate";
+import { Section } from "@/components/Section";
+
 import type { Route } from "./+types/index";
+
 import { changelog } from "./changelog.codegen";
-import { Demo } from "./Demo";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01SourceCode from "./demos/Demo01.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "Select | Copy UI" },
     { name: "description", content: "The Select component from Copy UI." },
@@ -16,12 +18,11 @@ export function meta({}: Route.MetaArgs) {
 export default function SelectPage() {
   return (
     <ComponentTemplate component="Select">
-      <Section
-        title="Select"
-        demoAndCode={[<Demo />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="Select">
+        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }