+++ src/components/Select/CHANGELOG.md
@@ -3,3 +3,5 @@
 - 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render
   prop.
 - 2025-10-03: Fix missing dependencies in useEffect hooks, etc.
+- 2025-10-11: Support `id` prop for better form integration. Make it be able to
+  uncontrolled. Support `defaultValue` prop. Code formatting improvements.